switching to diskcache implementation of task unique lock
why? celery_once does not work with chains. the errors are not
propagating, and sometimes there are some dangling locks.

so instead I implemented this locking mechanism via diskcache locks
and decorators.

disadvantage of this implementation: things on a local file
system. however, this can be fixed if one uses memcache locks.

ah yes, and the caching results are now via the decorators