fix: make Cassandra_Path json serialisable
why? celery works only with json serialisable objects.

previously we returned a Cassandra_Path which is a class inherited
from a string. celery easily serialised this to string, and this
become indistinguishable from a string

this made problems with remote workers, as get_locally method did not
retrieved remotely cached data.