fix: cluster points in route in metric coordinates
why?

I use metric coordinates in ssdp, and cluster points to be located in
degree coordinate boxes.

The problem is that the sampled raster in metric coordinates (that we
need for ssdp), looks like a parallelogram in the degree
coordinates. Hence some of the route point might end up outside the
metric box.