generate task queue in a separate task
get_Tasks_Queues stores a dictionary:
  (method, args) -> job_id

method defines a task to compute, and args define its argument.

job_id can be either a task generation task, or the task itself.

Tasks_Queues is implemented in redis, hence things are perishable
together with redis going down, which makes sense, as the queue has to
be regenerated, in case redis went down.

a job_id can be retrieved only once, and hence on retrieval a record
of such task must be deleted. otherwise, the task will be ever
pending.