catch exceptions from retrieve source and try another credentials
this covers the situation, when the counter stored in broker is
incorrect, so in case we did brake any limit on the request counts,
the task will try another credentials.

things will fail on when one tried the 3rd time the same credentials,
which *does not* guarantee that we tried every possible key, since
keys are requested from a circle by many processes.