implement call_cache_fn_results
this help to cache results of the `calls`, i.e. what is found in the
calls.py files.

The `call_cache_fn_results` adds an additional tasks to the query,
that take the resulting filename of the executed call, and assigns
this filename as a cache value for the call.