fix and reorganise paths for results_cache
fix: the tasks now are a task class (due to the bind=True,
80339e97a). This caused the first argument be a string like this:
  <@task: pvgrip.raster.tasks.sample_from_box of __main__ at
  0x7fbf3ddaa9d0>,
where 0x7fbf3ddaa9d0 is something that depends on the current compiled
version of pvgrip package, and not specific function. This caused
the cache not to be used, and filled the storage.

Further, the tasks results are now placed in a structured way, as in
that way it is easy to purge files that are no longer valid.