compute stdev of lidar points
how is it different from stat='stdev' for lidar data? it is not.

but in order to get 'stdev' for different window sizes, one requires
to run pdal for different pdal_resolution. this is pretty slow, and
might even require some extra space

this can be avoided, by running once pdal for 'stdev', 'mean',
'count', and the stdev for arbitrary windows can be computed
on-the-fly, as
  stdev, mean, count = ...          # are the sampled lidar points with
                                    # the corresponding statistics
  sum_in_region = ...               # sum in region operator
  xisq = count * (stdev^2 + mean^2)
  xi = count * mean
  count = sum_in_region(count)
  xi = sum_in_region(xi)
  xisq = sum_in_region(xisq)
  xisq / count - (xi/count)^2       # is the same things as stdev
                                    # lidar with different pdal_resolution

sum_in_region is a convolution