fix: avoid passing large data (route) through
in route computations the route was passed as dictionary

this can be pretty large, considering the dictionary contains all the
fields in the original dataset. The problem starts, when there are a
lot of cached results, hence there is a huge traffic surge when many
tasks are executed in a short period of time.

This makes celery workers to hit a timeout for the broker
reconnection, and things start to go down.