fix: avoid using epsg:3857; breaking changes!
=What happened?

Although epsg:3857 units are meters, 1m on equator equates to ~1.56m
in Germany. I was misled by "units", and didn't pay attention to the
accuracy.

=How was it fixed?

In order to fix this, I changed the behaviour of the 'mesh_type'
argument. Now, it accepts either epsg code of the coordinate system to
use, or the 'utm' string. Now a user is more flexible sampling his/her
raster. In procedures, where it is important that things are measured
in meters (irradiance/shadow/filter/route) the mesh_type='utm' is
enforced.

=How it was tested?

To test things I went to google maps, and measured length of different
buildings using haversine formula, and measured resulting lengths in
lidar data using GIMP's measure tool (number of pixels * step).

=Are there breaking changes?

Yes, in order to make sure that old sampled data is not used (pickle
files), the order of latitude and longitude is changed. Previously it
was longitude first. Now it is latitude first. This becomes consistent
with the 'box' argument as well. In case the old pickled data is used,
the resulting raster file will be scrambled.