fix: enforce correct dimension for the osm
I am still baffled of what is the reason for a wrong dimension
computations. For some reason, smrender computes correct height, if
one uses epsg:3857.

Correct here means same as in the `api/raster`.

With `utm` sampling, the width and height are different. In this
commit, I ensure that width and height are the same as in the
`api/raster`. However, the building positioning do not match anymore.